updated sqlalchemy 0.5 support with less magic and more meat

+++ tests/test_all.py
@@ -969,7 +969,8 @@ class TestSQLA(object):
             os.unlink('test_output/gh.sqlite')
 
 
-        metadata.drop_all()
+        if metadata.bind:
+            metadata.drop_all()
 
 
         # The SQLA tables
